+++ R/affirm_report.R
@@ -1,10 +1,13 @@
 #' Affirmation Report
 #'
 #' - `affirm_report_gt()` returns styled gt table summarizing results of affirmation session.
+#' - `affirm_report_excel()` returns excel file with one sheet per affirmation (excluding those with no errors)
 #' - `affirm_report_raw_data()` returns raw data used to generate summary in `affirm_report_gt()`
-#' - `affirm_export_excel()` returns excel file with one sheet per affirmation (excluding those with no errors)
 #'
 #' @inheritParams openxlsx::write.xlsx
+#' @param variable_labels logical indicating whether to add a row to exported
+#' data with the variable labels. If label does not exist, the variable name
+#' is printed. Default is `FALSE`
 #'
 #' @return gt table
 #' @name affirm_report
@@ -29,21 +32,22 @@ NULL
 
 #' @rdname affirm_report
 #' @export
-affirm_report_gt <- function() {
-  affirm_report_raw_data() |>
+affirm_report_gt <- function(variable_labels = FALSE) {
+  affirm_report_raw_data(variable_labels = variable_labels) |>
     dplyr::mutate(status_color = NA_character_, .before = 1L) |>
     dplyr::mutate(
-    csv_download_link =
-      mapply(
-        FUN = .as_csv_encoded_html_download_link,
-        # these two args are the ones being passed to FUN
-        .data$data,
-        paste0("extract_", dplyr::row_number(), ".csv"),
-        # additional mapply args
-        SIMPLIFY = TRUE,
-        USE.NAMES = FALSE
-      )
-  ) |>
+      csv_download_link =
+        mapply(
+          FUN = .as_csv_encoded_html_download_link,
+          # these two args are the ones being passed to FUN
+          .data$data,
+          paste0("extract_", dplyr::row_number(), ".csv"),
+          variable_labels,
+          # additional mapply args
+          SIMPLIFY = TRUE,
+          USE.NAMES = FALSE
+        )
+    ) |>
     dplyr::select(-"data") |>
     gt::gt() |>
     .affirm_report_gt_stylings()
@@ -52,9 +56,9 @@ affirm_report_gt <- function() {
 
 #' @rdname affirm_report
 #' @export
-affirm_export_excel <- function(file, overwrite = TRUE) {
+affirm_report_excel <- function(file, variable_labels = FALSE, overwrite = TRUE) {
   df_report <-
-    affirm_report_raw_data() |>
+    affirm_report_raw_data(variable_labels = variable_labels) |>
     dplyr::filter(.data$error_n > 0L) |>
     dplyr::mutate(
       label_no_specials = gsub(pattern = "[[:punct:]]", replacement = "", x = .data$label),
@@ -68,19 +72,27 @@ affirm_export_excel <- function(file, overwrite = TRUE) {
 
   df_report$data |>
     stats::setNames(df_report$label_no_specials) |>
-    openxlsx::write.xlsx(file = file, overwrite = overwrite)
+    openxlsx::write.xlsx(file = file, overwrite = overwrite, colNames = !variable_labels)
 }
 
 #' @rdname affirm_report
 #' @export
-affirm_report_raw_data <- function() {
+affirm_report_raw_data <- function(variable_labels = FALSE) {
   .check_affirm_initialized()
-  get(x = "df_affirmations", envir = env_affirm_logs) |>
+  df_report <-
+    get(x = "df_affirmations", envir = env_affirm_logs) |>
     dplyr::mutate(
       error_rate = .data$error_n / .data$total_n,
       .after = "total_n"
     ) |>
     dplyr::arrange(.data$error_n == 0L, .data$priority, dplyr::desc(.data$error_rate))
+
+  # add the variable labels to the top row if requested
+  if (isTRUE(variable_labels)) {
+    df_report$data <- lapply(df_report$data, .add_row_of_column_labels)
+  }
+
+  df_report
 }